# api/oas-doc.yaml
openapi: 3.0.0
info:
  version: 1.3.0
  title: Innosoft API
  description: >-
    Service to expose innosoft database as a RESTFUL API following Open Api v3
    specification. 
  contact:
    name: Pingu Developers
    email: pingudevelop@gmail.com
    url: 'https://github.com/orgs/Pingu-Developers'
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
paths:
...
  /api/v1/messages/{roomId}:
    get:
      description: Returns all messages from a room
      operationId: findMessagesByRoomId
      parameters:
        - name: roomId
          in: path
          description: roomId of messages to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/message'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        '400':
          description: Invalid roomId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
    post:
      description: Post a new message to a room
      operationId: addMessageByRoomId
      parameters:
        - name: roomId
          in: path
          description: roomId of message to add
          required: true
          schema:
            type: string
      requestBody:
        x-name: message
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/message'
        description: Message to be added
        required: true
      responses:
        '201':
          description: Message posted
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorModel'
components:
  schemas:
    errorModel:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    message:
      type: object
      required:
        - messageText
        - messageDateTime
        - messageUser
      properties:
        messageId:
          type: string
        messageText:
          type: string
        messageDateTime:
          type: string
          format: date-time
        messageUser:
          type: object
          required:
            - userName
          properties:
            userId:
              type: string
            userName:
              type: string
            userAvatar:
              type: string